Logging / scheudler issue resolution
Changed awslogs configuration to only look for the webserver.log and the scheduler.log on the websched instance and the worker.log on the worker instance.

Added daily restart of the aws cloudwatch logging service at 2am every morning

Also added a restart for the scheduler every morning at 3 am.